Resolve configuration resetting concern

+++ src/main/java/com/drtshock/playervaults/PlayerVaults.java
@@ -311,7 +311,7 @@ public class PlayerVaults extends JavaPlugin {
 
     private void loadConfig() {
         File configYaml = new File(this.getDataFolder(), "config.yml");
-        if (configYaml.exists()) {
+        if (!(new File(this.getDataFolder(), "config.conf").exists()) && configYaml.exists()) {
             this.config.setFromConfig(this.getLogger(), this.getConfig());
             try {
                 Files.move(configYaml.toPath(), this.getDataFolder().toPath().resolve("old_unused_config.yml"));
